Fix peer race condition, details below.
This happened when remote dialed the local server quicker than the local
was able to add peer (from the mdns event). Current fix is to add
peer whenever it's possible (as early as possible) and check
if peer exists in the peer list when handling other events.

It's possible that different flow would help. Actor model maybe?